refactor: major record overhaul
There is no InactiveRecord struct, instead all records are born in
inactive state without both timestamp and formatted message. This
change was required to properly implement filtering system in a sync
logger.

However (and that’s mentioned) I think that such common filtering can
be an overhead for some cases and can be easily replaced with less
common but more performance-gained filtering such as severity
comparison.

Very likely that there will be more breaking work aimed to filtering
and records.